+++ frontend/src/components/Brands.js
@@ -91,15 +91,15 @@ const Brands = () => {
             <div className="iconincircle">
               <img src="images/main/image51.png"/>
             </div>
-            <h5>Яндекс.Карты</h5>
-            <p>55 млн чел./месяц</p>
+            <h5>2ГИС</h5>
+            <p>30 млн чел./месяц</p>
           </div>
           <div className="brandicon">
             <div className="iconincircle">
               <img src="/images/main/image52.png"/>
             </div>
-            <h5>Google карты</h5>
-            <p>30 млн чел./месяц</p>
+            <h5>Wildberries</h5>
+            <p>92 млн чел./месяц</p>
           </div>
           </div>
         </Carousel.Item>
@@ -109,15 +109,33 @@ const Brands = () => {
             <div className="iconincircle">
               <img src="images/main/image54.png"/>
             </div>
-            <h5>Яндекс.Карты</h5>
-            <p>55 млн чел./месяц</p>
+            <h5>Zoon</h5>
+            <p>24 млн чел./месяц</p>
           </div>
           <div className="brandicon">
             <div className="iconincircle">
               <img src="/images/main/image55.png"/>
             </div>
-            <h5>Google карты</h5>
-            <p>30 млн чел./месяц</p>
+            <h5>OZON</h5>
+            <p>80 млн чел./месяц</p>
+          </div>
+          </div>
+        </Carousel.Item>
+        <Carousel.Item>
+        <div className="slide">
+          <div className="brandicon">
+            <div className="iconincircle">
+              <img src="images/main/image54.png"/>
+            </div>
+            <h5>Yell</h5>
+            <p>3 млн чел./месяц</p>
+          </div>
+          <div className="brandicon">
+            <div className="iconincircle">
+              <img src="/images/main/image55.png"/>
+            </div>
+            <h5>Flamp</h5>
+            <p>50 млн чел./месяц</p>
           </div>
           </div>
         </Carousel.Item>